rewrite memory management and add tests
In zebu.c there are now two new structures to manage memory:

struct zz_blob is similar to the old zz_node_chunk and zz_string_chunk:
it allocates memory in blocks of at least ZZ_BLOB_SIZE bytes---only
when a block larger than this is requested will all memory be allocated
in a cingle block at once---, and is used for both nodes and strings.

struct zz_string_index keeps a record of all strings allocated by the
tree, so the same string won't be allocated twice. The current record
is in the form of a simple list, so an obvious update would be to sort
them for easier searching.