add support for locations
The zz_node struct has not a loc field of type struct zz_location,
where the location from which that particular node was parsed can be
written.

The zz_tree has a callback function called gen_loc and a field called
gen_loc_data that can be used to automatically assign locations to
nodes as they are created. This can be made by passing a FLEX scanner
as data to a function that assigns the line and column variables to the
location passed to the function.