fix size comparison when checking for space in blobs
The check that determines if there is space remaining in a blob doesn't
take into account that some blobs are bigger than ZZ_BLOB_SIZE; since
all operations are between unsigned integers, there is the possibility
of getting negative results that would be interpreted as huge numbers,
allowing for data to be added at the end of full blobs.

A simple change for an equivalent equation ensures that all numbers
involved in the check remain positive and fixes the problem.