chidren list no longer the same as node data
The list of children of each node is now separate from the node data.
This means inner nodes may hold data too: an operator can hold an
integer determining what particular operation it is, or it can be used
to tag nodes with user-defined data after the tree has been built.

The zz_inner function does not exist anymore; a similar result may be
achieved by creating an empty node with zz_null and then appending the
children with zz_append.