allow nodes of arbitrary size
When initializing a tree, the size of its nodes can be provided as an
argument; the tree will use this size when allocating nodes.

If the node size is greater than sizeof(struct zz_node)---it can't be
less than that---, the extra bytes are zeroed when the node is
constructed, expecting the user to initialize them later.

This may be used to add extra fields to nodes; see tests/locations.c
for a simple example in which a single integer number is added to the
node.

Deprecates the old location field, since it allows used to define their
own location data.

Also, tree has now an extra field called error that stores a function
pointer that can be used to report errors from nodes. Users can provide
their own functions that operate on their own node types.