remove pointer to tree from node
Nodes do not require a pointer to their trees anymore. For this to
work, the following changes happened:

Tokens are not integer but constant strings, that are directly
referenced by the nodes. They can still be used the same way by
declaring the string correctly:

	const char *TOK_FOO = "foo"

This means that the zz_tree_init function does require neither the
num_tokens nor the token_names arguments anymore.

The zz_match functions take now an additional argument in the form of
the error handling function.