add zz_foreach_child, rename zz_list_foreach
The zz_list_foreach and zz_list_reverse_foreach macros are now called
just zz_foreach and zz_reverse_foreach.

Added the zz_foreach_child and zz_reverse_foreach_child macros; they
operate on a similar way to zz_foreach and zz_reverse_foreach, but the
iterator is supposed to be a pointer to zz_node, and the iterable is
the parent node of the children to iterate.