faster memory management for zz_tree
The new memory manager on zz_tree allocates chunks of 8, 16, 32, 64,
128, 256, 512, or 1024 bytes. These chunks are allocated together in
memory blobs, each of which holds segments of the same size, greatly
easing finding a free chunk, but increasing the amount of memory
needed.

The check to determine if a string has already been allocated in the
zz_tree has been removed to further speed up allocation at the cost of
even more memory.