add "golden" versions of all test outputs
When make test is invoked, the output of every test <test>.log is
matched against a golden file <test>.log.gold; the test succeeds only
if both files are equal.

Changed the print test so the pointer object is always NULL, instead of
a value that can change between executions.