Add foreach_child macros, fix unref'ing of nodes
There are now four specialized foreach macros that operate on the child
list of zz_node objects.

The first and last child, and next and prev sibling functions do not
increment the reference counter of nodes anymore; the counter is only
incremented when the node is added to a list and decremented when
removed.