Dict are ref. counted; add lookup, delete functions
Each zz_dict object has now a reference counter to determine how many
times the string has been added to the dictionary.

The zz_dict_delete function can now be used to remove entries from a
dictionary. The strings are actually deallocated when the function is
called as many times as the zz_dict_insert was.

The zz_dict_lookup function can be used to retrieve a string without
incrementing its reference counter.