Strings are handled globally
The dictionary that indexes strings used by trees is now the sime for
ill trees. This allows different trees to share the strings they hold:
a situation that may happen more often than not. It also removes the
need of passing a pointer to the tree to a number of functions.